# platform/swaggo/swagger.yaml
basePath: /flighthours/api/v1
definitions:
...
  handlers.AirlineListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      airlines:
        items:
          $ref: '#/definitions/handlers.AirlineResponse'
        type: array
      total:
        type: integer
    type: object
  handlers.AirlineResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      airline_code:
        type: string
      airline_name:
        type: string
      id:
        type: string
      status:
        type: string
      uuid:
        type: string
    type: object
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@flighthours.com
    name: Flighthours API Support
    url: https://flighthours.com/support
  description: API RESTful para la plataforma Flighthours, implementada con arquitectura
    hexagonal y siguiendo los principios del Richardson Maturity Model (Nivel 2-3)
    con HATEOAS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flighthours Backend API
  version: "1.0"
paths:
  /airlines:
    get:
      description: Returns a list of all airlines with optional status filter
      parameters:
      - description: Filter by status (true for active, false for inactive)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AirlineListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all airlines
      tags:
      - Airlines
  /airlines/{id}:
    get:
      consumes:
      - application/json
      description: Returns airline information by ID (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Airline ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get airline by ID
      tags:
      - Airlines
  /airlines/{id}/activate:
    patch:
      consumes:
      - application/json
      description: Sets airline status to active (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Airline ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Activate airline
      tags:
      - Airlines
  /airlines/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Sets airline status to inactive (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Airline ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deactivate airline
      tags:
      - Airlines
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"